app/src/main/java/com/jameskbride/codemashcompanion/data/ConferenceRepository.kt
package com.jameskbride.codemashcompanion.data

import com.jameskbride.codemashcompanion.bus.*
import org.greenrobot.eventbus.EventBus
import org.greenrobot.eventbus.Subscribe
import org.greenrobot.eventbus.ThreadMode
import javax.inject.Inject

class ConferenceRepository @Inject constructor(private val conferenceDao: ConferenceDao, override val eventBus: EventBus): BusAware {

    @Subscribe(threadMode = ThreadMode.BACKGROUND)
    fun onSpeakersReceivedEvent(speakersReceivedEvent: SpeakersReceivedEvent) {
        conferenceDao.insertAll(speakersReceivedEvent.speakers)
        eventBus.post(SpeakersPersistedEvent())
    }

    @Subscribe(threadMode = ThreadMode.BACKGROUND)
    fun onSessionsReceivedEvent(sessionsReceivedEvent: SessionsReceivedEvent) {
        conferenceDao.insertAll(sessionsReceivedEvent.sessions)
        eventBus.post(ConferenceDataPersistedEvent())
    }
}